Move code from modules/blocks/api to api/controllers/blocks
